// ButorRevolutionWebAPI/src/frontend/web/projects/factory/src/app/admin/users/models/users.models.ts
export interface IUserNotificationViewModel {
    label: string;
    notifications: IUserNotificationModel[];
}
export interface IUserNotificationDataModel {
    notificationType: string[];
    eventTypeIds: number[];
}
export interface IUserNotificationModel {
    id?: number;
    isChecked?: boolean;
    name: string;
}